renames new slice to cache slice
