+++ extension/src/popup/App.tsx
@@ -10,7 +10,7 @@ import { reducer as auth } from "popup/ducks/accountServices";
 import { reducer as settings } from "popup/ducks/settings";
 import { reducer as transactionSubmission } from "popup/ducks/transactionSubmission";
 import { reducer as tokenPaymentSimulation } from "popup/ducks/token-payment";
-import { reducer as balances } from "popup/ducks/balances";
+import { reducer as cache } from "popup/ducks/cache";
 import { ErrorTracking } from "popup/components/ErrorTracking";
 import { AccountMismatch } from "popup/components/AccountMismatch";
 import { ErrorBoundary } from "./components/ErrorBoundary";
@@ -23,7 +23,7 @@ const rootReducer = combineReducers({
   settings,
   transactionSubmission,
   tokenPaymentSimulation,
-  balances,
+  cache,
 });
 export type AppState = ReturnType<typeof rootReducer>;
 export const store = configureStore({
