+++ extension/src/popup/__testHelpers__/index.tsx
@@ -8,7 +8,7 @@ import { Balances } from "@shared/api/types/backend-api";
 
 import { reducer as auth } from "popup/ducks/accountServices";
 import { reducer as settings } from "popup/ducks/settings";
-import { reducer as balances } from "popup/ducks/balances";
+import { reducer as cache } from "popup/ducks/cache";
 import { defaultBlockaidScanAssetResult } from "@shared/helpers/stellar";
 import {
   reducer as transactionSubmission,
@@ -29,7 +29,7 @@ const rootReducer = combineReducers({
   settings,
   transactionSubmission,
   tokenPaymentSimulation,
-  balances,
+  cache,
 });
 
 const makeDummyStore = (state: any) =>
